chore: update Channel/useSendFileMessageCallback to ts
